gpiocmds: MCU command for immediate software PWM update
update_digital_out_pwm_value follows a similar semantic as update_digital_out, but interfaces correctly with queued changes and the software PWM cycle itself to provide a means to alter the PWM value of a software controlled PWM pin with no delay.

Had to do some mildly awkward things to get the PRU build to fit - if there's a means to exclude this command from that build, it would probably be for the best (along with handling of its absence in Klippy) to exclude it and de-awkward this implementation.

Signed-off-by: Brian Bossé <penrif@gmail.com>