heaters: Suppress updates relative to value
A 5% change in a heater's value has a very different impact depending on how much of the heater's max is being utilized. Recognizing this, let's allow through updates that are larger than 5% of the current value instead of 5% of the max value. This greatly increases the accuracy of heaters working on the low end of their power capability.

Signed-off-by: Brian Bossé <penrif@gmail.com>